+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed "Do-or-die" effect to no longer include "Sleep" challenge.
 - Changed "Enable Spawn Protection" option to also increase respawn immunity time if spawn is deemed too dangerous.
+- Fixed "Craft Item" challenge only choosing Campfire.
 - Fixed "Spawn Fake Guardian" effect so that it no longer spawns projectiles. This happens in Fargo's Mod Eternity Mode
   and Calamity Mod Revengeance Mode.